oxide/src/main/kotlin/io/github/oxidefrp/oxide/Signal.kt
package io.github.oxidefrp.oxide

import io.github.oxidefrp.oxide.signal.ConstantSignalVertex
import io.github.oxidefrp.oxide.signal.MapSignalVertex
import io.github.oxidefrp.oxide.signal.SignalVertex
import io.github.oxidefrp.oxide.signal.SourceSignalVertex

abstract class Signal<out A> {
    companion object {
        fun <A> constant(value: A): Signal<A> =
            object : Signal<A>() {
                override val vertex: SignalVertex<A> =
                    ConstantSignalVertex(value = value)
            }

        fun <A> source(sampleExternal: () -> A): Signal<A> =
            object : Signal<A>() {
                override val vertex: SignalVertex<A> =
                    SourceSignalVertex(sampleExternal = sampleExternal)
            }
    }

    internal abstract val vertex: SignalVertex<A>

    fun <B> map(transform: (A) -> B): Signal<B> =
        object : Signal<B>() {
            override val vertex: SignalVertex<B> =
                MapSignalVertex(
                    source = this@Signal.vertex,
                    transform = transform,
                )
        }
}
